+++ src/components/app/controls/main-controls/index.jsx
@@ -1,16 +1,32 @@
 import React, { useState, useEffect } from 'react';
 import propTypes from 'prop-types';
+import Popover from 'react-tiny-popover';
 import isMobile from '@config/is-mobile';
+import TimerConfigContainer from '@containers/timer-config.container';
 import {
   faPlay,
   faRandom,
   faStepBackward,
   faStepForward,
   faStop,
+  faHourglassStart,
+  faHourglassHalf,
+  faHourglassEnd,
 } from '@fortawesome/free-solid-svg-icons';
 import ControlButtonComponent from '../control-button';
-import ButtonSpacerComponent from './button-spacer';
 import './main-controls.scss';
+
+const getTimerIcon = ({ lastDurationsMS, remainingMS }) => {
+  const [lastDurationMS] = lastDurationsMS;
+  const percentComplete = remainingMS / lastDurationMS;
+
+  if (percentComplete >= 0.66) {
+    return faHourglassStart;
+  } else if (percentComplete >= 0.33) {
+    return faHourglassHalf;
+  }
+  return faHourglassEnd;
+};
 
 const makePrimaryButton = (faIcon, onClick) =>
   function PrimaryButtonComponent() {
@@ -33,28 +49,51 @@ const MainControlsComponent = ({
   onStopClick,
   onPlayClick,
   isRecordingGenerationInProgress,
+  timer,
 }) => {
   if (isRecordingGenerationInProgress) {
     return <div className="main-controls">Generating recording...</div>;
   }
 
+  const [isPopoverOpen, setIsPopoverOpen] = useState(true);
+
+  const toggleIsPopoverOpen = () => {
+    setIsPopoverOpen(val => !val);
+  };
   const PrimaryButtonComponent = isPlaying
     ? makePrimaryButton(faStop, onStopClick)
     : makePrimaryButton(faPlay, onPlayClick);
+
   return (
     <div className="main-controls">
-      {!isMobile && <ButtonSpacerComponent />}
+      <Popover
+        isOpen={isPopoverOpen}
+        content={<TimerConfigContainer done={() => setIsPopoverOpen(false)} />}
+        onClickOutside={toggleIsPopoverOpen}
+        contentLocation={({ targetRect, popoverRect }) => ({
+          top: targetRect.top - popoverRect.height - 5,
+          left: targetRect.x + targetRect.width / 2 - popoverRect.width / 2,
+        })}
+      >
+        <ControlButtonComponent
+          faIcon={getTimerIcon(timer)}
+          onClick={toggleIsPopoverOpen}
+          isActive={timer.remainingMS > 0}
+        />
+      </Popover>
       <ControlButtonComponent
         faIcon={faStepBackward}
         onClick={onPreviousClick}
       />
       <PrimaryButtonComponent />
       <ControlButtonComponent faIcon={faStepForward} onClick={onNextClick} />
-      <ControlButtonComponent
-        faIcon={faRandom}
-        onClick={isShuffleActive ? disableShuffle : enableShuffle}
-        isActive={isShuffleActive}
-      />
+      {!isMobile && (
+        <ControlButtonComponent
+          faIcon={faRandom}
+          onClick={isShuffleActive ? disableShuffle : enableShuffle}
+          isActive={isShuffleActive}
+        />
+      )}
     </div>
   );
 };