+++ src/store/actions/creators/update-timer.creator.js
@@ -1,0 +1,5 @@
+import UPDATE_TIMER from '../types/update-timer.type';
+
+const updateTimer = deltaMS => ({ type: UPDATE_TIMER, payload: deltaMS });
+
+export default updateTimer;